+++ lib/trailblazer/activity/dsl/linear/state.rb
@@ -76,7 +76,7 @@ module Trailblazer
           def self.update_sequence_for(type, task, options={}, sequence:, **kws, &block)
             step_options = invoke_normalizer_for(type, task, options, sequence: sequence, **kws, &block)
 
-            _sequence = Linear::Sequence.apply_adds(sequence, step_options[:adds])
+            _sequence = Activity::Adds.apply_adds(sequence, step_options[:adds])
           end
 
           def self.invoke_normalizer_for(type, task, options, normalizers:, normalizer_options:, sequence:, &block)